replaced example, scripts, test folders; fixed algorithm bugs in forced_aligner.py; vosk_align.py: get_result(args) was extracted from main() for testing; test_align.py: passing of args for testing has been changed to get_result(args)
