--- python/vosk/aligner/__init__.py
@@ -1,2 +1,0 @@
-from .forced_aligner import ForcedAligner
-from .transcription import Transcription
