+++ python/vosk/aligner/examples/cats.txt
@@ -1,0 +1,1 @@
+There was in this singular caravan little boy with no father or mother, but only a tiny kitten to cherish. The plague had not to him, yet had left him furry thing to mitigate his sorrow; and when one, one can find great relief in the lively antics of. So the boy whom the dark people called than he wept as he sat playing with his steps of an oddly painted wagon.
