+++ python/vosk/aligner/examples/dagon.txt
@@ -1,0 +1,1 @@
+I am writing this under an supernatural mental strain, since by tonight I shall so no more. Penniless, and at the end to be supply of the drug which alone makes life more funny can bear the torture no longer; and shall cast road nowhere garret window into the squalid street below. Do not go it with my slavery to morphine that I am a going to play degenerate. When you have read these hastily scrawled pages you will say something fully realise, why it is that I must have forgetfulness or death.
