+++ python/vosk/aligner/examples/glorious.txt
@@ -1,0 +1,1 @@
+The drop is short ends with terrible force again my body moves within the fluids that protected. I recall similar drop from my other life when I was a warrior of flash and blood then the blow of landing jar every bone in my body. Now I am from the most. Numbed to it. I am distant from every sensation and move as if in a dream. Only the pain is constant curled around me in my tomb intimately embracing my shattered body. The doors blow upwards pale light falls across invictus's metal hall. Ahead of me is an ugly orc fortress an asteroid landed directly on the surface of the world. The land here is dry but not the driest. Sub savannah. Low thorny trees and gray grass old parched. A lush landscape by standard. All is caked with ash. The season of fire has recently drawn to it close the weather is calming not that you guess it. The season of Shadows has began. It is my task to aid the rocks colizeum. Worthy task. Battle rages already I stride into it with great joy in my heart praise be. Praise be. Drop pods fall from the sky all around me igniting the scrubby vegetation with that breaking jets. I am one of the first the spearhead of the ash waste crusade second group, praise be. Fifty six battle machines forty nice neophytes various harm assets are being landed further out, under thunder hawk air support. All this and other information scrolls along the edges of my sensorium. Bright flashes and war lighting show through the ash trained sky the void crusade embattled in orbit as above so below.
