+++ python/vosk/aligner/examples/model/graph/disambig_tid.int
@@ -1,0 +1,17 @@
+10015
+10016
+10017
+10018
+10019
+10020
+10021
+10022
+10023
+10024
+10025
+10026
+10027
+10028
+10029
+10030
+10031
