+++ python/vosk/aligner/examples/model/ivector/online_cmvn.conf
@@ -1,0 +1,1 @@
+# configuration file for apply-cmvn-online, used in the script ../local/run_online_decoding.sh
