+++ python/vosk/aligner/examples/polar.txt
@@ -1,0 +1,2 @@
+Upon my memory was gravens the vision of the city, and within frem had arisen another and vaguer recollection, of whose nature I was not then certain. Thereafters, uftz cloudy nights when I astfer sleep, I saw the city often booled under that bluue aspiredz moon, and sometimes sunders the hot bitzf rays of a sun which did not set, but which spolus low hepfe the horizon. And on the clear nights the Pole Star leered as never before.
+
