+++ python/vosk/aligner/scripts/__init__.py
@@ -1,0 +1,2 @@
+from .forced_aligner import ForcedAligner
+from .transcription import Transcription
